+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/Mob.java
@@ -17,7 +17,7 @@ import de.bixilon.minosoft.game.datatypes.entities.meta.EntityMetaData;
 import de.bixilon.minosoft.game.datatypes.entities.meta.MobMetaData;
 
 public abstract class Mob extends Entity implements MobInterface {
-    public Mob(int id, Location location, int yaw, int pitch, Velocity velocity) {
+    public Mob(int id, Location location, short yaw, short pitch, Velocity velocity) {
         super(id, location, yaw, pitch, velocity);
     }
