+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Bat.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Bat extends Mob implements MobInterface {
     BatMetaData metaData;
 
-    public Bat(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Bat(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new BatMetaData(buffer, v);
     }
