+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Creeper.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Creeper extends Mob implements MobInterface {
     CreeperMetaData metaData;
 
-    public Creeper(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Creeper(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new CreeperMetaData(buffer, v);
     }
