+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/EnderDragon.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class EnderDragon extends Mob implements MobInterface {
     MobMetaData metaData;
 
-    public EnderDragon(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public EnderDragon(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new MobMetaData(buffer, v);
     }
