+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/EnderMan.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class EnderMan extends Mob implements MobInterface {
     EndermanMetaData metaData;
 
-    public EnderMan(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public EnderMan(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new EndermanMetaData(buffer, v);
     }
