+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Ghast.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Ghast extends Mob implements MobInterface {
     GhastMetaData metaData;
 
-    public Ghast(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Ghast(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new GhastMetaData(buffer, v);
     }
