+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/GiantZombie.java
@@ -20,7 +20,7 @@ import de.bixilon.minosoft.protocol.protocol.InByteBuffer;
 import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 
 public class GiantZombie extends Zombie {
-    public GiantZombie(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public GiantZombie(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity, buffer, v);
     }
