+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/IronGolem.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class IronGolem extends Mob implements MobInterface {
     MobMetaData metaData;
 
-    public IronGolem(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public IronGolem(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new MobMetaData(buffer, v);
     }
