+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/MagmaCube.java
@@ -21,7 +21,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 
 public class MagmaCube extends Slime {
 
-    public MagmaCube(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public MagmaCube(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity, buffer, v);
     }
