+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Ocelot.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Ocelot extends Mob implements MobInterface {
     OcelotMetaData metaData;
 
-    public Ocelot(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Ocelot(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new OcelotMetaData(buffer, v);
     }
