+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/OtherPlayer.java
@@ -29,7 +29,7 @@ public class OtherPlayer extends Mob implements MobInterface {
     HumanMetaData metaData;
     Pose status = Pose.STANDING;
 
-    public OtherPlayer(int id, String name, UUID uuid, PlayerPropertyData[] properties, Location location, Velocity velocity, int yaw, int pitch, short currentItem, HumanMetaData metaData) {
+    public OtherPlayer(int id, String name, UUID uuid, PlayerPropertyData[] properties, Location location, Velocity velocity, short yaw, short pitch, short currentItem, HumanMetaData metaData) {
         super(id, location, yaw, pitch, velocity);
         this.name = name;
         this.uuid = uuid;
