+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Pig.java
@@ -23,7 +23,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Pig extends Mob implements MobInterface {
     PigMetaData metaData;
 
-    public Pig(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Pig(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new PigMetaData(buffer, v);
     }
