+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Slime.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Slime extends Mob implements MobInterface {
     SlimeMetaData metaData;
 
-    public Slime(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Slime(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new SlimeMetaData(buffer, v);
     }
