+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Villager.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Villager extends Mob implements MobInterface {
     VillagerMetaData metaData;
 
-    public Villager(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Villager(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new VillagerMetaData(buffer, v);
     }
