+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/mob/Witch.java
@@ -22,7 +22,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Witch extends Mob implements MobInterface {
     WitchMetaData metaData;
 
-    public Witch(int id, Location location, int yaw, int pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
+    public Witch(int id, Location location, short yaw, short pitch, Velocity velocity, InByteBuffer buffer, ProtocolVersion v) {
         super(id, location, yaw, pitch, velocity);
         this.metaData = new WitchMetaData(buffer, v);
     }
