+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/objects/Arrow.java
@@ -25,10 +25,9 @@ public class Arrow extends EntityObject implements ObjectInterface {
     ArrowMetaData metaData;
     int shooter;
 
-    public Arrow(int id, Location location, int yaw, int pitch, int additionalInt, ProtocolVersion v) {
+    public Arrow(int id, Location location, short yaw, short pitch, int additionalInt, ProtocolVersion v) {
         super(id, location, yaw, pitch, null);
         // objects do not spawn with metadata... reading additional info from the following int
-        // tnt does not have any additional info
         this.shooter = additionalInt;
     }
