+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/objects/Boat.java
@@ -24,7 +24,7 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class Boat extends EntityObject implements ObjectInterface {
     BoatMetaData metaData;
 
-    public Boat(int id, Location location, int yaw, int pitch, int additionalInt, ProtocolVersion v) {
+    public Boat(int id, Location location, short yaw, short pitch, int additionalInt, ProtocolVersion v) {
         super(id, location, yaw, pitch, null);
         // objects do not spawn with metadata... reading additional info from the following int
         // boat does not have any additional info
