+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/objects/FallingDragonEgg.java
@@ -23,10 +23,9 @@ import de.bixilon.minosoft.protocol.protocol.ProtocolVersion;
 public class FallingDragonEgg extends EntityObject implements ObjectInterface {
     EntityMetaData metaData;
 
-    public FallingDragonEgg(int id, Location location, int yaw, int pitch, int additionalInt, ProtocolVersion v) {
+    public FallingDragonEgg(int id, Location location, short yaw, short pitch, int additionalInt, ProtocolVersion v) {
         super(id, location, yaw, pitch, null);
         // objects do not spawn with metadata... reading additional info from the following int
-        // tnt does not have any additional info
     }
 
     @Override
