+++ src/main/java/de/bixilon/minosoft/game/datatypes/entities/objects/Snowball.java
@@ -24,10 +24,9 @@ public class Snowball extends EntityObject implements ObjectInterface {
     EntityMetaData metaData;
     int thrower;
 
-    public Snowball(int id, Location location, int yaw, int pitch, int additionalInt, ProtocolVersion v) {
+    public Snowball(int id, Location location, short yaw, short pitch, int additionalInt, ProtocolVersion v) {
         super(id, location, yaw, pitch, null);
         // objects do not spawn with metadata... reading additional info from the following int
-        // tnt does not have any additional info
         this.thrower = additionalInt;
     }
